feat: implement countdown for staking contract migration and refactor related components
